server/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
whitenoise==5.1.0
psycopg2_binary==2.8.6
gunicorn==20.0.4